back/requirements.txt
Flask
ariadne
google-cloud-storage
